BT: Pairing fails when it coincides with role switch with 3rd device
Authentication complete after role switch with a device after
role switch caused initialization of sdp and finally change
in bond state machine. This becomes a problem a bonding activity
is also in progress.

Added condition to ensure initialization of sdp only when
authentication complete has arrived from the same device for
which bonding was initialized.

Change-Id: I7b213ac73ba5107f5da7d11305cd349e4d310062
CRs-Fixed: 1105371