ESR11-E7716: Disable scans while setting local RPA
This change is required to disable any ongoing classic
advertisement or any scan before setting the RPA for the
DUT, and reenable them after the Set RPA command.
If a direct connection is ongoing, the RPA is not set.
A background connection is suspended and resumed for
sending the "Set RPA" HCI command.

Change-Id: I7b1f68cfce066a321742a9b5de05d6aa5a5224c4