A2DP: Fix for incoming connection failure in dual a2dp mode.
In dual a2dp mode, when first stream is connected, all the SEPs
for that particular stream are marked busy for avdtp discovery
respose for second device as only one remote device is can be
connected per stream. Certain headsets try to connect to SEPs
that is marked busy. Check if one of the SEPs in the cluster of
the acceptor sep is already connected. If yes then reject the
setconfiguration cmd.

Change-Id: I93114e4330370f1b19cd4f67f0d367ce6bc6cd00