LE: new_encryption_key shall be reset once ACL link dropped
In device A, the bonded device record is removed and reloaded
because user disable/enable BT. In device B, the bonded device
record is not changed. This will cause the value of new_encryption_key
is set to 0 in device A, but the value is still is 1 in device B.
new_encryption_key shall be reset once ACL link dropped.

CRs-Fixed: 2026485
Change-Id: I0ebbdfa7db967a131066413d5734dccaca3e8058