Bluetooth: Update AAC blacklist
Disable AAC codec for certain BT headset. This particular HS is
sending flow off after receving few audio packets with AAC format.
So DUT cannot sending any l2cap msgs till the remote flow on, and
remote goes to bad state and flow is not turned on and the
connection is dropped. Blacklisting HS in interest for AAC.

Change-Id: If1ea4b3ac64f37d97487bbfbd2500de609cd94d0