Clear remote suspend flag and send proper start.
When AVRCP pausing, clear remote suspend flag and send
the protocol start when we are sending response to the
event playstatus changed as part of notification response
as well as sending the response to the get play status.
Also send the vendor specific start to the SOC when we
are clearing the remote suspend flag for AVDTP suspend.

CRs-Fixed: 2021829
Change-Id: I6e1c7fe5ab0fcad23f752410342b0ddec5d126de