AVRCP: Create rsp type array wrt to different PDUs
For back to back query of same PDUs from remote side
without DUT's response in-between, we are ending up
overwriting the response type to zero for the second
responses onwards of the same PDU due to which we end
up sending non desired responses for subsequent PDU
commands. Hence, creating PDU specific response array
for response type as well.

CRs-Fixed: 2009225
Change-Id: I339d7ae6d41116ca8fbab1463ef875128c4d0d93